Added location validation logic to containerapp env.
